feat(settings): add golden tests for LlmSettingsPage
- Created settings_page_golden_test.dart in the correct directory.
- Generated golden reference images for all tabs and loading state.
- Cleaned up redundant old test file.
- Adjusted viewport size to prevent overflows in tests.
